JSON-only config, calibration int16 fix, unified popup buttons, process edit mode
Config
- Replace the fragile binary .cfg with JSON-only storage (FilMachine.json).
  New dependency-free mini JSON parser (main/mini_json.c/.h) builds on both the
  board and the SDL simulator (no cJSON). readConfigFile/readSettingsOnly parse
  JSON; writeConfigFile serialises it with a crash-safe .tmp + rename.
- genFilMachineCFG.py emits JSON only; COMMANDS.md updated for the .json copy.
- Repurpose the old chem/wb container-ml bytes as calibrated fill times
  (chemCalibFillSecs / wbCalibFillSecs); tank capacity is now a fallback only.

Temperature calibration
- Fix offset overflow: tempCalibOffset/chemCalibOffset were int8 (+/-12.7C) and
  wrapped (sign flip) on larger corrections; widen to int16 and clamp to +/-30C.
- Add a Reset button inside the calibration popup (Cancel/Reset/Set) and drop
  the hard-to-discover long-press reset.
- Real temperatures in Self Check and the process final step; bang-bang
  thermostat during development; DS18B20 read robustness.

Machine I/O
- Port B: 6 chemical level sensors (B0-B5) + 2 heater MOSFETs (B6/B7).
- Hall-gated START ("Tank in position"); draining honours discardAfterProc
  (WASTE vs origin container); smart fill/clean/drain time model from capacity
  and MIN/MAX sensors with calibrated fill.
- Maintenance Fill (water bath + chemistry), target-aware: WB inlet uses the
  flow meter, chemistry uses the WB->C1 pump; stop on the MAX sensor.
- Defer the boot valve self-test to the sysMan task so the display comes up
  immediately instead of blocking ~4s.

UI
- Unify all tune/confirm popup buttons (Fill, Calibration, Pump/Rotation Speed,
  Volume) to the Clean popup style: msgbox size, bottom-left/right placement,
  montserrat_30 font, red Cancel / green Set. Label reads "Set" (was "SET").
- Reset to Defaults now asks for confirmation (Cancel/OK) before resetting.
- Process detail read-only vs edit mode: existing processes open read-only with
  an Edit button, new processes open editable. Edit unlocks name/temp/tolerance/
  temp-control/film-type/preferred and the step list; Save returns to read-only.
  Read-only fades the editable controls and hides the cursor; steps cannot be
  swiped/reordered/deleted and open their detail read-only on tap.

Tests
- Update settings/persistence/utilities tests for the renamed calibration fields.

+++ tests/test_persistence.c
@@ -396,7 +396,7 @@ static void test_tempCalibOffset_persistence(void)
     struct machineSettings *s = &gui.page.settings.settingsParams;
 
     /* Save original */
-    int8_t orig_offset = s->tempCalibOffset;
+    int16_t orig_offset = s->tempCalibOffset;
 
     /* Set to a known value (-15 = -1.5°C) */
     s->tempCalibOffset = -15;
